infra: modificações estruturais nas pastas
criação da pasta sql e migrations, mudanças no router e no load de variáveis de ambiente
