+++ localhost.key
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDlSd45o0krEYhY
+6t/SplmPB2HSqClmR8CAp4LmUzqzJFFOT4XxBCJcGg2xmOdJF1ZPKOJ22zeypu2m
+Uj64I13+AlSoZvQTJzsASWMdgAgU4H9nHILUODb8WgHaz3wVYjbK9WHp8SruiNP4
+FQSEmcBF1WYf5p6e9IHQafiM2TyVIE3VTotM+yIo3DkYp6PpcjrJKuwPHygBOpHK
+Nh3ZqdkFEGli262sYLeDdquut+ItEtkhGpHnAzubi15GVfP503OLvGUO6+wrjBP3
+Bfqvhgf0rQrme875wxFyos1F6/fUo9d62j/bkUD0puIq+yPzsvvAKPVTJv13di5z
+NAmDwzkjAgMBAAECggEBAM3i5oka9jiWJbiyLwkYFpM4Lb7bAVG2XEUqJYBse+kg
+U2HeRoeFDtVlD6bJheU772RlcfUII5toCT2v3GG8YEoWDa6pw8DvGSvIuhdjP95/
+N+d+0nt9fuJacA/q/6Cz8D+O170ZsVM/yhaZyNTIur36QRmkHArfwzFB9s2Cw8Oi
+6yZoC26d0fF9tMNM5Um7hA2U+JA2wAVc4OZ/4GmFL+qKErV0VhokY96mDfUw3s5I
+CERfZS6TaMZF+7sR/2Lj9J4E31l8TIgAVXr7k2z5xbKpUaBehL0Xal6uUp3xeSlw
+YuLLXZJ084HNtpfyLclwWuBwmqBpK800TYgYwBm8RQECgYEA8kxikBDdsvK1iYGD
+6VUcakYFiXlgXqa634yn+fpGN0HGTvLlFTx6B7+rjQPnbX0ZAoAcA0bwz+Noj8/h
+uzmSJ9t9t5epMAU1ATBaRl7hrVCx5N5R+SQWkifImDj0JjtSZVwo5Cg+8oL22RuP
+8z5s8nsD04Wnz6k02an9r+bTl6UCgYEA8kElrJ+/DN2MR3W5D0Hr9TkeZedA36S1
+irD0Jet5y1Eqp5gpxeT/Va9tGdDSmJs0WJfponiR1PyPqC0r/3JI/Srp6ZwrUnlA
+QSbtWNHEkZKnd5CD4SEj51KKXeyD5hWWcDoN7by0+kO+i8AId9yJq+3QGTy0uFuM
+hn+Lgi5IcycCgYBCR0DC1co/NAMSjRoPTpfKPkMvesMwwPznP6uHRr3ESbq8R9cA
+Ox3KVtTPaxz4JzsbIv4Gq3Yd97/Xbu75buMbAgBeox7Fovb/z0EydKvY4/T1fRcB
+Rqzh5/4IHbQzbRC55z7MQnINlmN9uftjd1UH8zx1tN7VgByT2dsqLcL9fQKBgDdy
+YcUY6wluy+SvNSQg2njOEf8Q5vBJKqQLzFXBh6bZdQHW+fipaSebd0UpXoo/O8Pr
+QTZGklpFVDwR2rgk2D/ZyqAFdoTx65r95eUmXe9TubDMCuNESotvgOvubmIDHuOg
+5f876dA+ISarVUTZ3UjUrYSZOGPZ0FgYaT3qifqRAoGBAJf9Z37vIX3PcqTGlw3Q
+Iy9CFYcuxK6187pBJ0pCAWrb+teAAhE7z+eHe0mRL4V6UreAa+QTz5eyJQuK1Z5C
+bOLlFflC4Ryu5r0/Gv6EDj/orkJfv3sn0uF8tyTq7z2k7bTv0QyDmX0HijqcuNKN
+nOgtYyCfqa1+Bsa+Eiy7rEqB
+-----END PRIVATE KEY-----
