+++ src/routes/routes_config.go
@@ -1,0 +1,10 @@
+package routes
+
+import "net/http"
+
+type Route struct {
+	URI                string
+	Metodo             string
+	Funcao             func(http.ResponseWriter, *http.Request)
+	RequerAutenticacao bool
+}
